+++ include/quicksr/genetic/selection/rank_selection.hpp
@@ -27,12 +27,12 @@ namespace qsr {
  */
 class RankSelection : public BaseSelection {
 public:
-    constexpr RankSelection(float sp) : sp(sp) {}
+    constexpr RankSelection(double sp) : sp(sp) {}
 
     std::shared_ptr<BaseSelector> get_selector(int npopulation) noexcept override;
 
 private:
-    float sp; // Selection pressure parameter
+    double sp; // Selection pressure parameter
 };
 
 }
